+++ Example/ReproExample/ReproExample/AppDelegate.h
@@ -1,0 +1,17 @@
+//
+//  AppDelegate.h
+//  ReproExample
+//
+//  Created by Masahiro Hayashi on 10/15/14.
+//  Copyright (c) 2014 Repro Inc. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+
+@end
+