+++ Example/ReproExample/ReproExample/FirstViewController.m
@@ -1,0 +1,32 @@
+//
+//  FirstViewController.m
+//  ReproExample
+//
+//  Created by Masahiro Hayashi on 10/15/14.
+//  Copyright (c) 2014 Repro Inc. All rights reserved.
+//
+
+#import "FirstViewController.h"
+#import <Repro/Repro.h>
+
+@interface FirstViewController ()
+@property (weak, nonatomic) IBOutlet UITextField *idField;
+@property (weak, nonatomic) IBOutlet UITextField *passwordField;
+@end
+
+@implementation FirstViewController
+
+- (void)viewDidLoad {
+    [super viewDidLoad];
+    // Do any additional setup after loading the view, typically from a nib.
+
+    [Repro mask:_passwordField];
+    [Repro mask:_idField];
+}
+
+- (void)didReceiveMemoryWarning {
+    [super didReceiveMemoryWarning];
+    // Dispose of any resources that can be recreated.
+}
+
+@end